+++ itables/javascript.py
@@ -10,6 +10,7 @@ from base64 import b64encode
 
 import numpy as np
 import pandas as pd
+import pandas.io.formats.style as pd_style
 
 try:
     import polars as pl
@@ -28,7 +29,18 @@ from .utils import read_package_file
 logging.basicConfig()
 logger = logging.getLogger(__name__)
 
+_OPTIONS_NOT_AVAILABLE_WITH_TO_HTML = {
+    "footer",
+    "column_filters",
+    "showIndex",
+    "maxBytes",
+    "maxRows",
+    "maxColumns",
+    "warn_on_unexpected_types",
+    "warn_on_int_to_str_conversion",
+}
 _ORIGINAL_DATAFRAME_REPR_HTML = pd.DataFrame._repr_html_
+_ORIGINAL_DATAFRAME_STYLE_REPR_HTML = pd_style.Styler._repr_html_
 _ORIGINAL_POLARS_DATAFRAME_REPR_HTML = pl.DataFrame._repr_html_
 _CONNECTED = True
 
@@ -80,10 +92,12 @@ def init_notebook_mode(
     if all_interactive:
         pd.DataFrame._repr_html_ = _datatables_repr_
         pd.Series._repr_html_ = _datatables_repr_
+        pd_style.Styler._repr_html_ = _datatables_repr_
         pl.DataFrame._repr_html_ = _datatables_repr_
         pl.Series._repr_html_ = _datatables_repr_
     else:
         pd.DataFrame._repr_html_ = _ORIGINAL_DATAFRAME_REPR_HTML
+        pd_style.Styler._repr_html_ = _ORIGINAL_DATAFRAME_STYLE_REPR_HTML
         pl.DataFrame._repr_html_ = _ORIGINAL_POLARS_DATAFRAME_REPR_HTML
         if hasattr(pd.Series, "_repr_html_"):
             del pd.Series._repr_html_
@@ -232,13 +246,29 @@ class JavascriptCode(str):
     pass
 
 
-def _datatables_repr_(df=None, tableId=None, **kwargs):
-    return to_html_datatable(df, tableId, connected=_CONNECTED, **kwargs)
+def _datatables_repr_(df):
+    return to_html_datatable(df, connected=_CONNECTED)
 
 
 def to_html_datatable(
-    df=None, caption=None, tableId=None, connected=True, import_jquery=True, **kwargs
+    df=None,
+    caption=None,
+    tableId=None,
+    connected=True,
+    import_jquery=True,
+    use_to_html=False,
+    **kwargs
 ):
+    if use_to_html or isinstance(df, pd_style.Styler):
+        return to_html_datatable_using_to_html(
+            df=df,
+            caption=caption,
+            tableId=tableId,
+            connected=connected,
+            import_jquery=import_jquery,
+            **kwargs
+        )
+
     """Return the HTML representation of the given dataframe as an interactive datatable"""
     # Default options
     for option in dir(opt):
@@ -414,6 +444,122 @@ def to_html_datatable(
     return output
 
 
+def to_html_datatable_using_to_html(
+    df=None, caption=None, tableId=None, connected=True, import_jquery=True, **kwargs
+):
+    """Return the HTML representation of the given dataframe as an interactive datatable,
+    using df.to_html() rather than the underlying dataframe data."""
+    options_not_available = set(kwargs).intersection(
+        _OPTIONS_NOT_AVAILABLE_WITH_TO_HTML
+    )
+    if options_not_available:
+        raise TypeError(
+            "These options are not available when using df.to_html: {}".format(
+                options_not_available
+            )
+        )
+
+    # Default options
+    for option in dir(opt):
+        if (
+            option not in _OPTIONS_NOT_AVAILABLE_WITH_TO_HTML
+            and option not in kwargs
+            and not option.startswith("__")
+            and option not in ["read_package_file"]
+        ):
+            kwargs[option] = getattr(opt, option)
+
+    for name, value in kwargs.items():
+        if value is None:
+            raise ValueError(
+                "Please don't pass an option with a value equal to None ('{}=None')".format(
+                    name
+                )
+            )
+
+    # These options are used here, not in DataTable
+    css = kwargs.pop("css")
+    classes = kwargs.pop("classes")
+    style = kwargs.pop("style")
+    tags = kwargs.pop("tags")
+
+    if caption is not None:
+        tags = '{}<caption style="white-space: nowrap; overflow: hidden">{}</caption>'.format(
+            tags, caption
+        )
+
+    eval_functions = kwargs.pop("eval_functions", None)
+    pre_dt_code = kwargs.pop("pre_dt_code")
+
+    if "dom" not in kwargs and _df_fits_in_one_page(df, kwargs):
+        kwargs["dom"] = "t"
+
+    # Load the HTML template
+    if connected:
+        output = read_package_file("html/datatables_template_connected.html")
+    else:
+        output = read_package_file("html/datatables_template.html")
+
+    if not import_jquery:
+        assert (
+            connected
+        ), "In the offline mode, jQuery is imported through init_notebook_mode"
+        output = replace_value(
+            output, "    import 'https://code.jquery.com/jquery-3.6.0.min.js';\n", ""
+        )
+
+    if isinstance(df, pd_style.Styler):
+        assert tableId is None, "Styler objects already have an uuid"
+        tableId = "T_" + df.uuid
+        table_html = df.to_html()
+    else:
+        tableId = tableId or str(uuid.uuid4())
+        table_html = df.to_html(table_id=tableId)
+
+    if style:
+        style = 'style="{}"'.format(style)
+    else:
+        style = ""
+
+    table_html = replace_value(
+        table_html,
+        '<table id="{}">'.format(tableId),
+        """<table id="{tableId}" class="{classes}"{style}>{tags}""".format(
+            tableId=tableId, classes=classes, style=style, tags=tags
+        ),
+    )
+
+    output = replace_value(
+        output,
+        '<table id="table_id"><thead><tr><th>A</th></tr></thead></table>',
+        table_html,
+    )
+    output = replace_value(output, "#table_id", "#{}".format(tableId))
+    output = replace_value(
+        output,
+        "<style></style>",
+        "<style>{}</style>".format(css),
+    )
+
+    # Export the DT args to JSON
+    dt_args = json_dumps(kwargs, eval_functions)
+
+    output = replace_value(
+        output, "let dt_args = {};", "let dt_args = {};".format(dt_args)
+    )
+    output = replace_value(
+        output,
+        "// [pre-dt-code]",
+        pre_dt_code.replace("#table_id", "#{}".format(tableId)),
+    )
+
+    # No data since we pass the html table
+    output = replace_value(output, 'dt_args["data"] = data;', "")
+    output = replace_value(output, "const data = [];", "")
+
+    return output
+
+
 def _column_count_in_header(table_header):
     return max(line.count("</th>") for line in table_header.split("</tr>"))
 
@@ -433,7 +579,7 @@ def _min_rows(kwargs):
 
 def _df_fits_in_one_page(df, kwargs):
     """Display just the table (not the search box, etc...) if the rows fit on one 'page'"""
-    return len(df) <= _min_rows(kwargs)
+    return len(df.index) <= _min_rows(kwargs)
 
 
 def safe_reset_index(df):